refactor(scripts): convert to direct imports
